pyproject.toml
[project]
name = "experience-study-skill"
version = "0.1.0"
description = "Unified skill-based actuarial copilot for experience studies."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.1",
    "openai>=2.28.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "plotly>=6.0.0",
    "pyarrow>=21.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.2",
    "scipy>=1.13.1",
    "streamlit>=1.50.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

